+++ Sources/Geometria/3D/Protocols/Plane/Plane3Type.swift
@@ -1,0 +1,4 @@
+/// A plane that represents a flat 2D surface in 3D space.
+public protocol Plane3Type: PlaneType where Vector: Vector3Type {
+    
+}